+++ pyaction/datetime_/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+"""
+pyaction.datetime_
+~~~~~~~~~~~~~~~~~~
+
+The datetime module supplies classes for manipulating dates and times. While
+date and time arithmetic is supported, the focus of the implementation is on
+efficient attribute extraction for output formatting and manipulation.
+
+"""
+
+from .objects import *
+
+#-----------------------------------------------------------------------------